pyproject.toml
[project]
name = "paperpulse"
version = "0.1.0"
description = "Paper monitoring, analysis and innovative IDEA recommendation system"
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "click>=8.0",
    "arxiv>=2.1",
    "requests>=2.28",
    "feedparser>=6.0",
    "PyMuPDF>=1.23",
    "openai>=1.0",
    "schedule>=1.2",
    "python-dateutil>=2.8",
    "tqdm>=4.65",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
mineru = ["mineru>=0.6"]
dev = ["pytest>=8.0", "pytest-cov>=5.0"]

[project.scripts]
paperpulse = "paperpulse.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/paperpulse"]